requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.1
Django==4.0.3
django-rest-knox==4.2.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
